Add special prepend logic for PATH variable
