+++ Cargo.toml
@@ -33,7 +33,7 @@ toml = {version = "0.7.5", features = ["preserve_order"]}
 assert_cmd = {version = "2.0.11", default-features = false, features = ["color-auto"]}
 pretty_assertions = "1.4.0"
 rstest = {version = "0.17.0", default-features = false}
-rstest_reuse = "0.5.0"
+rstest_reuse = "0.6.0"
 serde_test = "1.0.165"
 
 [package.metadata.release]
